+++ phpstan-bootstrap.php
@@ -5,18 +5,18 @@ if (!defined('ABSPATH')) {
     define('ABSPATH', __DIR__ . '/');
 }
 
-if (!defined('NABA_HDWP_PLUGIN_NAME')) {
-    define('NABA_HDWP_PLUGIN_NAME', 'hd-plugin-wordpress');
+if (!defined('NOVA_STATS_PLUGIN_NAME')) {
+    define('NOVA_STATS_PLUGIN_NAME', 'hd-plugin-wordpress');
 }
 
-if (!defined('NABA_HDWP_PLUGIN_URL')) {
-    define('NABA_HDWP_PLUGIN_URL', 'https://example.com/wp-content/plugins/hd-plugin-wordpress/');
+if (!defined('NOVA_STATS_PLUGIN_URL')) {
+    define('NOVA_STATS_PLUGIN_URL', 'https://example.com/wp-content/plugins/hd-plugin-wordpress/');
 }
 
-if (!defined('NABA_HDWP_PLUGIN_PATH')) {
-    define('NABA_HDWP_PLUGIN_PATH', __DIR__ . '/');
+if (!defined('NOVA_STATS_PLUGIN_PATH')) {
+    define('NOVA_STATS_PLUGIN_PATH', __DIR__ . '/');
 }
 
-if (!defined('NABA_HDWP_VERSION')) {
-    define('NABA_HDWP_VERSION', '0.0.1');
+if (!defined('NOVA_STATS_VERSION')) {
+    define('NOVA_STATS_VERSION', '0.0.1');
 }